+++ packages/partitura_core/lib/partitura_core.dart
@@ -28,6 +28,8 @@ export 'src/layout/page_layout.dart';
 export 'src/layout/score_layout.dart';
 export 'src/layout/staff_system.dart';
 export 'src/layout/tab_layout.dart';
+export 'src/mei/mei_reader.dart';
+export 'src/mei/mei_writer.dart';
 export 'src/midi/midi_reader.dart';
 export 'src/midi/midi_writer.dart';
 export 'src/model/element.dart';